Removed RemoteMetadataManager and moved RemoteDataFolder to main
